requirements.txt
bcrypt==3.2.2
blinker==1.5
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
PyJWT==2.4.0
SQLAlchemy==1.4.39
Werkzeug==2.2.0
WTForms==3.0.1
zipp==3.8.1
